+++ src/main/java/net/sybyline/scarlet/ScarletStaffList.java
@@ -5,12 +5,16 @@ import java.io.Reader;
 import java.io.Writer;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import io.github.vrchatapi.model.User;
 
 import net.sybyline.scarlet.util.MiscUtils;
 
@@ -24,16 +28,56 @@ public class ScarletStaffList
         this.scarlet = scarlet;
         this.staffListFile = staffListFile;
         this.staffUserIds = Collections.newSetFromMap(new ConcurrentHashMap<>());
+        this.staffNames = new ConcurrentHashMap<>();
         this.load();
     }
 
     final Scarlet scarlet;
     final File staffListFile;
     final Set<String> staffUserIds;
+    final Map<String, String> staffNames;
 
     public String[] getStaffIds()
     {
         return this.staffUserIds.toArray(new String[this.staffUserIds.size()]);
+    }
+
+    public String getStaffName(String userId)
+    {
+        if (!this.isStaffId(userId))
+            return null;
+        String name = this.staffNames.get(userId);
+        if (name == null && this.scarlet.vrc != null)
+        {
+            User user = this.scarlet.vrc.getUser(userId);
+            if (user != null)
+            {
+                name = user.getDisplayName();
+            }
+        }
+        if (name == null)
+        {
+            name = userId;
+        }
+        return name;
+    }
+
+    public Map<String, String> getStaffNames()
+    {
+        Map<String, String> ret = new HashMap<>();
+        for (String userId : this.getStaffIds())
+        {
+            ret.put(userId, this.getStaffName(userId));
+        }
+        return ret;
+    }
+
+    public void populateStaffNames()
+    {
+        for (String userId : this.getStaffIds())
+        {
+            this.getStaffName(userId);
+        }
     }
 
     public boolean isStaffId(String userId)